fix: update skeleton file structure (#96451)
