+++ app/[tenant]/[workspace]/(subapps)/forum/common/ui/components/compose-post/index.ts
@@ -1,0 +1,1 @@
+export * from './compose-post';
