+++ app/[tenant]/[workspace]/(subapps)/forum/common/ui/components/hero/hero.tsx
@@ -3,7 +3,7 @@ import {useCallback, useState} from 'react';
 
 // ---- CORE IMPORTS ---- //
 import {i18n} from '@/locale';
-import {HeroSearch, Search} from '@/ui/components';
+import {HeroSearch, Search, Skeleton} from '@/ui/components';
 import {useWorkspace} from '@/app/[tenant]/[workspace]/workspace-context';
 import {useSearchParams} from '@/ui/hooks';
 import {
@@ -122,5 +122,21 @@ export function Hero() {
       }
       renderSearch={!selectedGroup && renderSearch}
     />
+  );
+}
+
+export function HeroSkeleton() {
+  return (
+    <div className="relative overflow-hidden flex-col lg:w-auto w-full h-[300px] lg:h-[353px] flex items-center justify-center bg-white mt-4">
+      <div className="flex flex-col gap-4 w-full items-center ">
+        <Skeleton className="w-20 h-20" />
+        <Skeleton className="w-96 h-8" />
+        <div className="flex flex-col gap-2 w-full items-center justify-center md:max-w-screen-sm lg:max-w-screen-md">
+          <Skeleton className="w-full h-4" />
+          <Skeleton className="w-5/6 h-4" />
+        </div>
+        <Skeleton className="w-3/5 h-12 md:max-w-screen-sm lg:max-w-screen-md" />
+      </div>
+    </div>
   );
 }
