+++ app/[tenant]/[workspace]/(subapps)/forum/common/ui/components/index.ts
@@ -1,7 +1,10 @@
 export {Search} from './search';
 export {GroupActionList} from './group-action-list';
-export {GroupNotification} from './group-notification';
-export {NavMenu} from './nav-menu';
+export {
+  GroupNotification,
+  ForumNotificationSkeleton,
+} from './group-notification';
+export {NavMenu, NavMenuSkeleton} from './nav-menu';
 export {UploadPost} from './upload-post';
 export {CreatePost} from './create-post';
 export {ImageUploader} from './image-uploader';
@@ -9,8 +12,14 @@ export {FileUploader} from './file-uploader';
 export {FileDesc} from './file-desc';
 export {Tabs} from './tabs';
 export {MediaContent} from './media-content';
-export {ThreadList} from './thread-list';
-export {Thread, ThreadHeader, ThreadBody, ThreadFooter} from './thread';
+export {ThreadList, ThreadListSkeleton} from './thread-list';
+export {
+  Thread,
+  ThreadHeader,
+  ThreadBody,
+  ThreadFooter,
+  ThreadSkeleton,
+} from './thread';
 export {ImageGallery} from './image-gallery';
 export {ThreadPopup} from './thread-popup';
 export {InfiniteScroll} from './infinite-scroll';
@@ -19,5 +28,6 @@ export {FilePreviewer} from './file-previewer';
 export {ImagePreviewer} from './image-previewer';
 export {ImageViewer} from './image-viewer';
 export {NotificationHeader} from './notification-header';
-export {Hero} from './hero';
-export {GroupControls} from './group-controls';
+export {Hero, HeroSkeleton} from './hero';
+export {GroupControls, GroupControlsSkeleton} from './group-controls';
+export {ComposePost} from './compose-post';
