+++ app/[tenant]/[workspace]/(subapps)/forum/common/ui/components/thread-list/thread-list.tsx
@@ -10,7 +10,10 @@ import {URL_PARAMS} from '@/constants';
 import {SORT_BY_OPTIONS} from '@/comments';
 
 // ---- LOCAL IMPORTS ---- //
-import {InfiniteScroll} from '@/subapps/forum/common/ui/components';
+import {
+  InfiniteScroll,
+  ThreadSkeleton,
+} from '@/subapps/forum/common/ui/components';
 import {Post} from '@/subapps/forum/common/types/forum';
 
 export const ThreadList = ({
@@ -58,3 +61,14 @@ export const ThreadList = ({
 };
 
 export default ThreadList;
+
+export function ThreadListSkeleton({postCount = 3}: {postCount?: number}) {
+  const showImageView = Math.floor(Math.random() * postCount);
+  return (
+    <div>
+      {[...Array(postCount)].map((_, i) => (
+        <ThreadSkeleton key={i} galleryPriview={i === showImageView} />
+      ))}
+    </div>
+  );
+}
