+++ app/[tenant]/[workspace]/(subapps)/forum/common/ui/components/thread/thread.tsx
@@ -7,6 +7,7 @@ import {
   ThreadHeader,
 } from '@/subapps/forum/common/ui/components';
 import type {Post} from '@/subapps/forum/common/types/forum';
+import {Skeleton} from '@/ui/components';
 
 interface ThreadProps {
   post?: Post;
@@ -69,3 +70,54 @@ export const Thread = ({
 };
 
 export default Thread;
+
+export function ThreadSkeleton({
+  galleryPriview = false,
+}: {
+  galleryPriview?: boolean;
+}) {
+  return (
+    <div className="flex flex-col gap-4 p-4 bg-white rounded-md mt-4">
+      <div>
+        <div className="flex gap-4 ">
+          <Skeleton className="rounded-full h-8 w-8" />;
+          <Skeleton className="h-5 w-32" />
+        </div>
+        <Skeleton className="w-full h-[1px] my-2" />
+      </div>
+      <div className="flex items-center justify-between">
+        <div className="flex items-center gap-4">
+          <Skeleton className="rounded-full h-10 w-10" />
+          <div className="flex flex-col gap-2">
+            <Skeleton className="h-5 w-32" />
+            <Skeleton className="h-3 w-24" />
+          </div>
+        </div>
+      </div>
+      <div className="flex flex-col gap-2">
+        <Skeleton className="h-7 w-3/4" />
+        <div className="space-y-2">
+          <Skeleton className="h-4 w-full" />
+          <Skeleton className="h-4 w-full" />
+          <Skeleton className="h-4 w-4/5" />
+        </div>
+        <div className="flex justify-end">
+          <Skeleton className="h-4 w-24" />
+        </div>
+      </div>
+      {galleryPriview && (
+        <div className="grid grid-cols-3 gap-2">
+          <Skeleton className="h-40 w-full" />
+          <Skeleton className="h-40 w-full" />
+          <Skeleton className="h-40 w-full" />
+        </div>
+      )}
+
+      <Skeleton className="w-full h-[1px] my-2" />
+      <div className="flex items-center gap-[0.625rem]">
+        <Skeleton className="w-full h-8" />
+        <Skeleton className="w-32 h-8" />
+      </div>
+    </div>
+  );
+}
